Address feedback and move script out of PR
