+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Math/AABB.cpp
@@ -6,7 +6,7 @@
  *
  */
 
-#include "AABBFunctions.h"
+#include "AABB.h"
 
 namespace ScriptCanvas
 {
